Move RCS settings to RemoteClusterSettings class
Relocates all Settings from RemoteClusterService to a
RemoteClusterSettings class.  Note that a
RemoteClusterSettingsTests class already exists that tests some of
these settings. This is one part in a series of refactors to use a
LinkedProjectConfig data class in RemoteClusterService and its related
classes.

Relates: ES-12656, ES-12569
