+++ server/src/main/java/org/elasticsearch/transport/SniffConnectionStrategy.java
@@ -537,9 +537,9 @@ public class SniffConnectionStrategy extends RemoteConnectionStrategy {
 
     // Default visibility for tests
     static Predicate<DiscoveryNode> getNodePredicate(Settings settings) {
-        if (RemoteClusterService.REMOTE_NODE_ATTRIBUTE.exists(settings)) {
+        if (RemoteClusterSettings.REMOTE_NODE_ATTRIBUTE.exists(settings)) {
             // nodes can be tagged with node.attr.remote_gateway: true to allow a node to be a gateway node for cross cluster search
-            String attribute = RemoteClusterService.REMOTE_NODE_ATTRIBUTE.get(settings);
+            String attribute = RemoteClusterSettings.REMOTE_NODE_ATTRIBUTE.get(settings);
             return DEFAULT_NODE_PREDICATE.and((node) -> Booleans.parseBoolean(node.getAttributes().getOrDefault(attribute, "false")));
         }
         return DEFAULT_NODE_PREDICATE;
